requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
absl-py=2.1.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
cfgv=3.4.0=pypi_0
charset-normalizer=3.4.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd8ed1ab_1
distlib=0.3.9=pypi_0
docformatter=1.7.5=pypi_0
einops=0.8.0=pypi_0
etils=1.11.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
ffmpeg=1.4=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.55.3=pypi_0
fsspec=2024.10.0=pypi_0
glfw=2.8.0=pypi_0
hydra-core=1.3.2=pypi_0
identify=2.6.3=pypi_0
imageio=2.36.1=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
importlib-resources=6.4.5=pypi_0
iniconfig=2.0.0=pypi_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.30.0=pyh707e725_0
jedi=0.19.2=pyhd8ed1ab_1
joblib=1.4.2=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h237132a_0
libcxx=19.1.5=ha82da77_0
libedit=3.1.20191231=hc8eb9b7_2
libffi=3.4.4=hca03da5_1
libsodium=1.0.20=h99b78c6_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mediapy=1.2.2=pypi_0
moviepy=2.1.1=pypi_0
mpmath=1.3.0=pypi_0
mujoco=3.2.6=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
nodeenv=1.9.1=pypi_0
numpy=2.2.0=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=pyhd8ed1ab_2
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=pypi_0
pip=24.2=py310hca03da5_0
platformdirs=4.3.6=pyhd8ed1ab_1
pluggy=1.5.0=pypi_0
pre-commit=4.0.1=pypi_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.48=pyha770c72_1
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.18.0=pyhd8ed1ab_1
pygsp=0.5.1=pypi_0
pyopengl=3.1.7=pypi_0
pyparsing=3.2.0=pypi_0
pytest=8.3.4=pypi_0
python=3.10.16=hb885b13_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py310h313beb8_0
readline=8.2=h1a28f6b_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py310hca03da5_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.45.3=h80987f9_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h6ba3021_0
tomli=2.2.1=pypi_0
tornado=6.4.2=py310h80987f9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=h04d1e81_0
untokenize=0.1.1=pypi_0
virtualenv=20.28.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h18a0788_1
